+++ Wellnest/Wellnest/Feature/ScheduleInput/Manual/Model/AlarmRule.swift
