+++ Wellnest/Wellnest/Feature/ScheduleInput/Manual/View/ScheduleItemView.swift
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct ScheduleRowView: View {
+struct ScheduleItemView: View {
     let schedule: ScheduleItem
 
     var body: some View {
