+++ Wellnest/Wellnest/Resource/Assets.xcassets/Color/accentCardYellowColor.colorset/Contents.json
